requirements.txt
arrow==0.15.4
asgiref==3.2.3
blessed==1.16.1
certifi==2019.11.28
chardet==3.0.4
dj-database-url==0.5.0
Django==3.0.3
django-cors-headers==3.2.0
django-picklefield==2.0
django-request-logging==0.7.1
django-rest-framework==0.1.0
djangorestframework==3.11.0
future==0.18.2
fylesdk==0.17.2
fyle-rest-auth==0.4.0
idna==2.8
psycopg2-binary==2.8.4
python-dateutil==2.8.1
pytz==2019.3
requests==2.22.0
requests-cache==0.5.2
six==1.13.0
sqlparse==0.3.0
urllib3==1.25.7
wcwidth==0.1.8
gspread
oauth2client==4.1.3
google-api-python-client